Let Hetzner decide which datacenter to choose
We are running quite often into an error like `we are unable to provision servers for this location, try with a different location or try later (resource_unavailable)`, the Hetzner Cloud support told me that we should be able to not provide any datacenter and Hetzner Cloud will automatically decide which datacenter could be taken for the new instance.

With this patch I have made the datacenter option optionally. With my
current tests it seems to be running fine as far as I can say.