Failed to ping server fix
The server will still be in a running state after ping failure, and the autoscaler will keep trying to reprovision it. The stopped time is set, but the state is not updated. This puts it in a stopped state, similar to https://github.com/drone/autoscaler/blob/master/engine/collect.go#L108-L109 and https://github.com/drone/autoscaler/blob/master/engine/reaper.go#L121-L122