Allow autoscaler to work with pre-existing static builders.
Use case: I have several reserved instances on equinix metal (nee Packet), but have credits on DigitalOcean that I use for overflow capacity.
Before this PR the autoscaler will overprovision, thankfully it doesn't impact the bill too much, but this PR would decrease our bill.
In theory, this could allow multiple instances of the autoscaler to run and have for ex 1-3 instances to run on provider 1, 4-5 with provider 2,
and anymore on provider 3 (way more complex use cases would also be possible).

Hopefully you find this PR useful, but if not then that's ok too as I could still use it for myself, but figured I'd contribute it so others
can use it as well.

Signed-off-by: Matti R <matti@mdranta.net>